添加各个模块

+++ Server/xiaomi_mall/src/main/java/com/xiaomi_mall/service/impl/UserDetailsServiceImpl.java
@@ -1,8 +1,8 @@
 package com.xiaomi_mall.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
-import com.xiaomi_mall.enity.User;
-import com.xiaomi_mall.enity.authentication.LoginUser;
+import com.xiaomi_mall.exception.enity.User;
+import com.xiaomi_mall.exception.enity.authentication.LoginUser;
 import com.xiaomi_mall.mapper.UserMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -30,7 +30,7 @@ public class UserDetailsServiceImpl implements UserDetailsService {
             throw new RuntimeException("用户名或密码错误");
         }
         //TODO 根据用户查询权限信息 添加到LoginUser中
-        Integer userId = user.getUserId();
+        Long userId = user.getUserId();
         List<String> permissionKeyList = userMapper.getUserType(userId);
 
         //封装成UserDetails对象返回
